chore: Undo `--all` and `--graph` wrap
chore: Removing dependency on `--all` in non-run commands

fix: Fixing integration with `runall` and `graph` in `run` command

fix: Restoring `render --all` functionality

fix: Fixing `--all` commands that got busted from not having the wrap

fix: Fixing `backend delete --all` tests

fix: Fixing `backend bootstrap --all`

fix: Use feature flag for controlling ss3encryption

fix: Fixing `TestAwsDeleteBackend` now that we're actually deleting
state for all units with `--all`

fix: Addressing review feedback
