+++ cli/commands/backend/bootstrap/cli.go
@@ -3,11 +3,9 @@ package bootstrap
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
 	"github.com/gruntwork-io/terragrunt/cli/flags"
 	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
-	"github.com/gruntwork-io/terragrunt/internal/runner/run"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 )
@@ -28,16 +26,17 @@ func NewFlags(opts *options.TerragruntOptions) cli.Flags {
 }
 
 func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
+	cmdFlags := NewFlags(opts)
+	cmdFlags = append(cmdFlags, shared.NewAllFlag(opts, nil), shared.NewFailFastFlag(opts))
+
 	cmd := &cli.Command{
 		Name:  CommandName,
 		Usage: "Bootstrap OpenTofu/Terraform backend infrastructure.",
-		Flags: NewFlags(opts),
+		Flags: cmdFlags,
 		Action: func(ctx context.Context, _ *cli.Context) error {
 			return Run(ctx, l, opts.OptionsFromContext(ctx))
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, true)
 
 	return cmd
 }
