+++ cli/commands/backend/delete/cli.go
@@ -3,11 +3,9 @@ package delete
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
 	"github.com/gruntwork-io/terragrunt/cli/flags"
 	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
-	"github.com/gruntwork-io/terragrunt/internal/runner/run"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 )
@@ -47,16 +45,17 @@ func NewFlags(l log.Logger, opts *options.TerragruntOptions, prefix flags.Prefix
 }
 
 func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
+	cmdFlags := NewFlags(l, opts, nil)
+	cmdFlags = append(cmdFlags, shared.NewAllFlag(opts, nil), shared.NewFailFastFlag(opts))
+
 	cmd := &cli.Command{
 		Name:  CommandName,
 		Usage: "Delete OpenTofu/Terraform state.",
-		Flags: NewFlags(l, opts, nil),
+		Flags: cmdFlags,
 		Action: func(ctx context.Context, _ *cli.Context) error {
 			return Run(ctx, l, opts.OptionsFromContext(ctx))
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, true)
 
 	return cmd
 }
