+++ cli/commands/hcl/format/cli.go
@@ -3,11 +3,9 @@ package format
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
 	"github.com/gruntwork-io/terragrunt/cli/flags"
 	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
-	"github.com/gruntwork-io/terragrunt/internal/runner/run"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 )
@@ -97,8 +95,6 @@ func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
 			return Run(ctx, l, opts.OptionsFromContext(ctx))
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, true)
 
 	return cmd
 }
