+++ cli/commands/hcl/validate/cli.go
@@ -3,12 +3,9 @@ package validate
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/graph"
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
 	"github.com/gruntwork-io/terragrunt/cli/flags"
 	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
-	"github.com/gruntwork-io/terragrunt/internal/runner/run"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 )
@@ -93,9 +90,6 @@ func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
 			return Run(ctx, l, opts.OptionsFromContext(ctx))
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, true)
-	cmd = graph.WrapCommand(l, opts, cmd, run.Run, true)
 
 	return cmd
 }
