+++ cli/commands/info/print/cli.go
@@ -3,10 +3,9 @@ package print
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
 	runcmd "github.com/gruntwork-io/terragrunt/cli/commands/run"
+	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
-	"github.com/gruntwork-io/terragrunt/internal/runner/run"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 )
@@ -16,17 +15,18 @@ const (
 )
 
 func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
+	cmdFlags := runcmd.NewFlags(l, opts, nil)
+	cmdFlags = append(cmdFlags, shared.NewAllFlag(opts, nil))
+
 	cmd := &cli.Command{
 		Name:      CommandName,
 		Usage:     "Print out a short description of Terragrunt context.",
 		UsageText: "terragrunt info print",
-		Flags:     runcmd.NewFlags(l, opts, nil),
+		Flags:     cmdFlags,
 		Action: func(ctx context.Context, _ *cli.Context) error {
-			return Run(ctx, l, opts)
+			return Run(ctx, l, opts.OptionsFromContext(ctx))
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, true)
 
 	return cmd
 }
