+++ cli/commands/render/cli.go
@@ -4,13 +4,11 @@ package render
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/graph"
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
 	runcmd "github.com/gruntwork-io/terragrunt/cli/commands/run"
 	"github.com/gruntwork-io/terragrunt/cli/flags"
+	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
 	"github.com/gruntwork-io/terragrunt/internal/errors"
-	"github.com/gruntwork-io/terragrunt/internal/runner/run"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 )
@@ -119,23 +117,23 @@ func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
 	prefix := flags.Prefix{CommandName}
 	renderOpts := NewOptions(opts)
 
+	cmdFlags := append(runcmd.NewFlags(l, opts, nil), NewFlags(renderOpts, prefix)...)
+	cmdFlags = append(cmdFlags, shared.NewAllFlag(opts, prefix))
+
 	cmd := &cli.Command{
 		Name:        CommandName,
 		Usage:       "Render the final terragrunt config, with all variables, includes, and functions resolved, in the specified format.",
 		Description: "This is useful for enforcing policies using static analysis tools like Open Policy Agent, or for debugging your terragrunt config.",
-		Flags:       append(runcmd.NewFlags(l, opts, nil), NewFlags(renderOpts, prefix)...),
+		Flags:       cmdFlags,
 		Action: func(ctx context.Context, _ *cli.Context) error {
 			tgOpts := opts.OptionsFromContext(ctx)
-			renderOpts := renderOpts.Clone()
-			renderOpts.TerragruntOptions = tgOpts
 
-			return Run(ctx, l, renderOpts)
+			clonedOpts := renderOpts.Clone()
+			clonedOpts.TerragruntOptions = tgOpts
+
+			return Run(ctx, l, clonedOpts)
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, true)
-	// TODO: For backward compatibility, remove after getting rid of the `render-json` command, as supporting the `graph` flag for the `render` command is pointless.
-	cmd = graph.WrapCommand(l, opts, cmd, run.Run, true)
 
 	return cmd
 }
