+++ cli/commands/run/cli.go
@@ -5,11 +5,12 @@ import (
 	"context"
 	"strings"
 
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/graph"
-	"github.com/gruntwork-io/terragrunt/cli/commands/common/runall"
+	"github.com/gruntwork-io/terragrunt/cli/flags/shared"
 	"github.com/gruntwork-io/terragrunt/internal/cli"
 	"github.com/gruntwork-io/terragrunt/internal/report"
+	"github.com/gruntwork-io/terragrunt/internal/runner/graph"
 	"github.com/gruntwork-io/terragrunt/internal/runner/run"
+	"github.com/gruntwork-io/terragrunt/internal/runner/runall"
 	"github.com/gruntwork-io/terragrunt/options"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 	"github.com/gruntwork-io/terragrunt/tf"
@@ -20,6 +21,9 @@ const (
 )
 
 func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
+	cmdFlags := NewFlags(l, opts, nil)
+	cmdFlags = append(cmdFlags, shared.NewAllFlag(opts, nil), shared.NewGraphFlag(opts, nil))
+
 	cmd := &cli.Command{
 		Name:        CommandName,
 		Usage:       "Run an OpenTofu/Terraform command.",
@@ -28,13 +32,20 @@ func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
 		Examples: []string{
 			"# Run a plan\nterragrunt run -- plan\n# Shortcut:\n# terragrunt plan",
 			"# Run output with -json flag\nterragrunt run -- output -json\n# Shortcut:\n# terragrunt output -json",
-			// TODO: Add this example back when we support `run --all` again.
-			//
-			// "# Run a plan against a Stack of configurations in the current directory\nterragrunt run --all -- plan",
 		},
-		Flags:       NewFlags(l, opts, nil),
+		Flags:       cmdFlags,
 		Subcommands: NewSubcommands(l, opts),
 		Action: func(ctx context.Context, cliCtx *cli.Context) error {
+			tgOpts := opts.OptionsFromContext(ctx)
+
+			if tgOpts.RunAll {
+				return runall.Run(ctx, l, tgOpts)
+			}
+
+			if tgOpts.Graph {
+				return graph.Run(ctx, l, tgOpts)
+			}
+
 			if len(cliCtx.Args()) == 0 {
 				return cli.ShowCommandHelp(ctx, cliCtx)
 			}
@@ -42,9 +53,6 @@ func NewCommand(l log.Logger, opts *options.TerragruntOptions) *cli.Command {
 			return Action(l, opts)(ctx, cliCtx)
 		},
 	}
-
-	cmd = runall.WrapCommand(l, opts, cmd, run.Run, false)
-	cmd = graph.WrapCommand(l, opts, cmd, run.Run, false)
 
 	return cmd
 }
@@ -78,7 +86,17 @@ func Action(l log.Logger, opts *options.TerragruntOptions) cli.ActionFunc {
 
 		r := report.NewReport().WithWorkingDir(opts.WorkingDir)
 
-		return run.Run(ctx, l, opts.OptionsFromContext(ctx), r)
+		tgOpts := opts.OptionsFromContext(ctx)
+
+		if tgOpts.RunAll {
+			return runall.Run(ctx, l, tgOpts)
+		}
+
+		if tgOpts.Graph {
+			return graph.Run(ctx, l, tgOpts)
+		}
+
+		return run.Run(ctx, l, tgOpts, r)
 	}
 }
