+++ cli/flags/shared/auth.go
@@ -1,0 +1,34 @@
+package shared
+
+import (
+	"github.com/gruntwork-io/terragrunt/cli/flags"
+	"github.com/gruntwork-io/terragrunt/internal/cli"
+	"github.com/gruntwork-io/terragrunt/options"
+)
+
+const (
+	AuthProviderCmdFlagName = "auth-provider-cmd"
+)
+
+// NewAuthProviderCmdFlag creates a flag for specifying the auth provider command.
+func NewAuthProviderCmdFlag(opts *options.TerragruntOptions, prefix flags.Prefix, commandName string) *flags.Flag {
+	tgPrefix := prefix.Prepend(flags.TgPrefix)
+	terragruntPrefix := prefix.Prepend(flags.TerragruntPrefix)
+
+	var terragruntPrefixControl flags.RegisterStrictControlsFunc
+	if commandName != "" {
+		terragruntPrefixControl = flags.StrictControlsByCommand(opts.StrictControls, commandName)
+	} else {
+		terragruntPrefixControl = flags.StrictControlsByGlobalFlags(opts.StrictControls)
+	}
+
+	return flags.NewFlag(
+		&cli.GenericFlag[string]{
+			Name:        AuthProviderCmdFlagName,
+			EnvVars:     tgPrefix.EnvVars(AuthProviderCmdFlagName),
+			Destination: &opts.AuthProviderCmd,
+			Usage:       "Run the provided command and arguments to authenticate Terragrunt dynamically when necessary.",
+		},
+		flags.WithDeprecatedEnvVars(terragruntPrefix.EnvVars("auth-provider-cmd"), terragruntPrefixControl),
+	)
+}
