+++ cli/flags/shared/config.go
@@ -1,0 +1,34 @@
+package shared
+
+import (
+	"github.com/gruntwork-io/terragrunt/cli/flags"
+	"github.com/gruntwork-io/terragrunt/internal/cli"
+	"github.com/gruntwork-io/terragrunt/options"
+)
+
+const (
+	ConfigFlagName = "config"
+)
+
+// NewConfigFlag creates a flag for specifying the Terragrunt config file path.
+func NewConfigFlag(opts *options.TerragruntOptions, prefix flags.Prefix, commandName string) *flags.Flag {
+	tgPrefix := prefix.Prepend(flags.TgPrefix)
+	terragruntPrefix := prefix.Prepend(flags.TerragruntPrefix)
+
+	var terragruntPrefixControl flags.RegisterStrictControlsFunc
+	if commandName != "" {
+		terragruntPrefixControl = flags.StrictControlsByCommand(opts.StrictControls, commandName)
+	} else {
+		terragruntPrefixControl = flags.StrictControlsByGlobalFlags(opts.StrictControls)
+	}
+
+	return flags.NewFlag(
+		&cli.GenericFlag[string]{
+			Name:        ConfigFlagName,
+			EnvVars:     tgPrefix.EnvVars(ConfigFlagName),
+			Destination: &opts.TerragruntConfigPath,
+			Usage:       "The path to the Terragrunt config file. Default is terragrunt.hcl.",
+		},
+		flags.WithDeprecatedEnvVars(terragruntPrefix.EnvVars("config"), terragruntPrefixControl),
+	)
+}
