+++ cli/flags/shared/doc.go
@@ -1,0 +1,5 @@
+// Package shared provides flags that are shared by multiple commands.
+//
+// This package is underutilized right now, as some more serious refactoring is needed to make sure all
+// shared flags use this package instead of reusing flags from other commands.
+package shared
