+++ cli/flags/shared/parallelism.go
@@ -1,0 +1,28 @@
+package shared
+
+import (
+	"github.com/gruntwork-io/terragrunt/cli/flags"
+	"github.com/gruntwork-io/terragrunt/internal/cli"
+	"github.com/gruntwork-io/terragrunt/options"
+)
+
+const (
+	ParallelismFlagName = "parallelism"
+)
+
+// NewParallelismFlag creates a flag for specifying parallelism level.
+func NewParallelismFlag(opts *options.TerragruntOptions) *flags.Flag {
+	tgPrefix := flags.Prefix{flags.TgPrefix}
+	terragruntPrefix := flags.Prefix{flags.TerragruntPrefix}
+	terragruntPrefixControl := flags.StrictControlsByGlobalFlags(opts.StrictControls)
+
+	return flags.NewFlag(
+		&cli.GenericFlag[int]{
+			Name:        ParallelismFlagName,
+			EnvVars:     tgPrefix.EnvVars(ParallelismFlagName),
+			Destination: &opts.Parallelism,
+			Usage:       "Parallelism for --all commands.",
+		},
+		flags.WithDeprecatedEnvVars(terragruntPrefix.EnvVars("parallelism"), terragruntPrefixControl),
+	)
+}
