+++ cli/flags/shared/tfpath.go
@@ -1,0 +1,32 @@
+package shared
+
+import (
+	"github.com/gruntwork-io/terragrunt/cli/flags"
+	"github.com/gruntwork-io/terragrunt/internal/cli"
+	"github.com/gruntwork-io/terragrunt/options"
+)
+
+const (
+	TFPathFlagName = "tf-path"
+)
+
+// NewTFPathFlag creates a flag for specifying the OpenTofu/Terraform binary path.
+func NewTFPathFlag(opts *options.TerragruntOptions) *flags.Flag {
+	tgPrefix := flags.Prefix{flags.TgPrefix}
+	terragruntPrefix := flags.Prefix{flags.TerragruntPrefix}
+	terragruntPrefixControl := flags.StrictControlsByGlobalFlags(opts.StrictControls)
+
+	return flags.NewFlag(
+		&cli.GenericFlag[string]{
+			Name:    TFPathFlagName,
+			EnvVars: tgPrefix.EnvVars(TFPathFlagName),
+			Usage:   "Path to the OpenTofu/Terraform binary. Default is tofu (on PATH).",
+			Setter: func(value string) error {
+				opts.TFPath = value
+				opts.TFPathExplicitlySet = true
+				return nil
+			},
+		},
+		flags.WithDeprecatedEnvVars(terragruntPrefix.EnvVars("tfpath"), terragruntPrefixControl),
+	)
+}
