+++ internal/runner/runall/runall.go
@@ -1,3 +1,4 @@
+// Package runall implements the logic for running commands across all units in a stack.
 package runall
 
 import (
@@ -31,14 +32,6 @@ var runAllDisabledCommands = map[string]string{
 	tf.CommandNameUntaint:     "terraform untaint should only be run against a single state representation to avoid using the wrong state address.",
 	tf.CommandNameConsole:     "terraform console requires stdin, which is shared across all instances of run --all when multiple modules run concurrently.",
 	tf.CommandNameForceUnlock: "lock IDs are unique per state representation and thus should not be run with run --all.",
-
-	// MAINTAINER'S NOTE: There are a few other commands that might not make sense, but we deliberately allow it for
-	// certain use cases that are documented here:
-	// - state          : Supporting `state` with run --all could be useful for a mass pull and push operation, which can
-	//                    be done en masse with the use of relative pathing.
-	// - login / logout : Supporting `login` with run --all could be useful when used in conjunction with mise and
-	//                    multi-terraform version setups, where multiple terraform versions need to be configured.
-	// - version        : Supporting `version` with run --all could be useful for sanity checking a multi-version setup.
 }
 
 func Run(ctx context.Context, l log.Logger, opts *options.TerragruntOptions) error {
