+++ test/integration_aws_test.go
@@ -129,9 +129,44 @@ func TestAwsBootstrapBackend(t *testing.T) {
 			}
 
 			commonConfigPath := filepath.Join(rootPath, "common.hcl")
-			helpers.CopyTerragruntConfigAndFillPlaceholders(t, commonConfigPath, commonConfigPath, s3BucketName, dynamoDBName, helpers.TerraformRemoteStateS3Region)
+			helpers.CopyTerragruntConfigAndFillPlaceholders(
+				t,
+				commonConfigPath,
+				commonConfigPath,
+				s3BucketName,
+				dynamoDBName,
+				helpers.TerraformRemoteStateS3Region,
+			)
 
-			stdout, stderr, err := helpers.RunTerragruntCommandWithOutput(t, "terragrunt "+tc.args+" --all --non-interactive --log-level debug --working-dir "+rootPath)
+			// Also replace placeholders in subdirectory config files that are discovered by --all
+			dualLockingConfigPath := filepath.Join(rootPath, "dual-locking", "terragrunt.hcl")
+			require.FileExists(t, dualLockingConfigPath)
+
+			helpers.CopyTerragruntConfigAndFillPlaceholders(
+				t,
+				dualLockingConfigPath,
+				dualLockingConfigPath,
+				s3BucketName,
+				dynamoDBName,
+				helpers.TerraformRemoteStateS3Region,
+			)
+
+			useLockfileConfigPath := filepath.Join(rootPath, "use-lockfile", "terragrunt.hcl")
+			require.FileExists(t, useLockfileConfigPath)
+
+			helpers.CopyTerragruntConfigAndFillPlaceholders(
+				t,
+				useLockfileConfigPath,
+				useLockfileConfigPath,
+				s3BucketName,
+				dynamoDBName,
+				helpers.TerraformRemoteStateS3Region,
+			)
+
+			stdout, stderr, err := helpers.RunTerragruntCommandWithOutput(
+				t,
+				"terragrunt "+tc.args+" --all --non-interactive --log-level debug --working-dir "+rootPath,
+			)
 			require.NoError(t, err)
 
 			tc.checkExpectedResultFn(t, stdout+stderr, s3BucketName, dynamoDBName)
@@ -244,16 +279,69 @@ func TestAwsBootstrapBackendWithoutVersioning(t *testing.T) {
 	commonConfigPath := filepath.Join(rootPath, "common.hcl")
 	helpers.CopyTerragruntConfigAndFillPlaceholders(t, commonConfigPath, commonConfigPath, s3BucketName, dynamoDBName, helpers.TerraformRemoteStateS3Region)
 
-	_, _, err := helpers.RunTerragruntCommandWithOutput(t, "terragrunt run --all --non-interactive --log-level debug --working-dir "+rootPath+" apply --backend-bootstrap --feature disable_versioning=true")
+	// Also replace placeholders in subdirectory config files that are discovered by --all
+	dualLockingConfigPath := filepath.Join(rootPath, "dual-locking", "terragrunt.hcl")
+	require.FileExists(t, dualLockingConfigPath)
+
+	helpers.CopyTerragruntConfigAndFillPlaceholders(
+		t,
+		dualLockingConfigPath,
+		dualLockingConfigPath,
+		s3BucketName,
+		dynamoDBName,
+		helpers.TerraformRemoteStateS3Region,
+	)
+	// Add skip_bucket_versioning to disable_versioning feature
+	contents, err := util.ReadFileAsString(dualLockingConfigPath)
+	require.NoError(t, err)
+	anchorText := "    enable_lock_table_ssencryption = feature.enable_lock_table_ssencryption.value"
+	require.Contains(t, contents, anchorText, "Expected anchor text not found in %s", dualLockingConfigPath)
+	newContents := strings.ReplaceAll(contents, anchorText, anchorText+"\n    skip_bucket_versioning         = true")
+	require.NotEqual(t, contents, newContents, "strings.ReplaceAll did not modify contents of %s", dualLockingConfigPath)
+	err = os.WriteFile(dualLockingConfigPath, []byte(newContents), 0644)
+	require.NoError(t, err)
+
+	useLockfileConfigPath := filepath.Join(rootPath, "use-lockfile", "terragrunt.hcl")
+	require.FileExists(t, useLockfileConfigPath)
+
+	helpers.CopyTerragruntConfigAndFillPlaceholders(
+		t,
+		useLockfileConfigPath,
+		useLockfileConfigPath,
+		s3BucketName,
+		dynamoDBName,
+		helpers.TerraformRemoteStateS3Region,
+	)
+	// Add skip_bucket_versioning for disable_versioning feature
+	contents, err = util.ReadFileAsString(useLockfileConfigPath)
+	require.NoError(t, err)
+	anchorText = "    use_lockfile = true"
+	require.Contains(t, contents, anchorText, "Expected anchor text not found in %s", useLockfileConfigPath)
+	newContents = strings.ReplaceAll(contents, anchorText, anchorText+"\n    skip_bucket_versioning = true")
+	require.NotEqual(t, contents, newContents, "strings.ReplaceAll did not modify contents of %s", useLockfileConfigPath)
+	err = os.WriteFile(useLockfileConfigPath, []byte(newContents), 0644)
+	require.NoError(t, err)
+
+	_, _, err = helpers.RunTerragruntCommandWithOutput(
+		t,
+		"terragrunt run --all --non-interactive --log-level debug --working-dir "+rootPath+" apply --backend-bootstrap --feature disable_versioning=true",
+	)
 	require.NoError(t, err)
 
 	validateS3BucketExistsAndIsTaggedAndVersioning(t, helpers.TerraformRemoteStateS3Region, s3BucketName, false, nil)
 	validateDynamoDBTableExistsAndIsTaggedAndIsSSEncrypted(t, helpers.TerraformRemoteStateS3Region, dynamoDBName, nil, false)
 
-	_, _, err = helpers.RunTerragruntCommandWithOutput(t, "terragrunt --non-interactive --log-level debug --working-dir "+rootPath+" backend delete --backend-bootstrap --feature disable_versioning=true --all")
-	require.NoError(t, err)
+	_, _, err = helpers.RunTerragruntCommandWithOutput(
+		t,
+		"terragrunt --non-interactive --log-level debug --working-dir "+rootPath+" backend delete --backend-bootstrap --feature disable_versioning=true --all",
+	)
+	require.Error(t, err)
+	assert.Contains(t, err.Error(), "backend delete for unit")
 
-	_, _, err = helpers.RunTerragruntCommandWithOutput(t, "terragrunt --non-interactive --log-level debug --working-dir "+rootPath+" backend delete --backend-bootstrap --feature disable_versioning=true --all --force")
+	_, _, err = helpers.RunTerragruntCommandWithOutput(
+		t,
+		"terragrunt --non-interactive --log-level debug --working-dir "+rootPath+" backend delete --backend-bootstrap --feature disable_versioning=true --all --force",
+	)
 	require.NoError(t, err)
 }
 
@@ -339,7 +427,39 @@ func TestAwsDeleteBackend(t *testing.T) {
 	}()
 
 	commonConfigPath := filepath.Join(rootPath, "common.hcl")
-	helpers.CopyTerragruntConfigAndFillPlaceholders(t, commonConfigPath, commonConfigPath, s3BucketName, dynamoDBName, helpers.TerraformRemoteStateS3Region)
+	helpers.CopyTerragruntConfigAndFillPlaceholders(
+		t,
+		commonConfigPath,
+		commonConfigPath,
+		s3BucketName,
+		dynamoDBName,
+		helpers.TerraformRemoteStateS3Region,
+	)
+
+	// Also replace placeholders in subdirectory config files that are discovered by --all
+	dualLockingConfigPath := filepath.Join(rootPath, "dual-locking", "terragrunt.hcl")
+	require.FileExists(t, dualLockingConfigPath)
+
+	helpers.CopyTerragruntConfigAndFillPlaceholders(
+		t,
+		dualLockingConfigPath,
+		dualLockingConfigPath,
+		s3BucketName,
+		dynamoDBName,
+		helpers.TerraformRemoteStateS3Region,
+	)
+
+	useLockfileConfigPath := filepath.Join(rootPath, "use-lockfile", "terragrunt.hcl")
+	require.FileExists(t, useLockfileConfigPath)
+
+	helpers.CopyTerragruntConfigAndFillPlaceholders(
+		t,
+		useLockfileConfigPath,
+		useLockfileConfigPath,
+		s3BucketName,
+		dynamoDBName,
+		helpers.TerraformRemoteStateS3Region,
+	)
 
 	_, _, err := helpers.RunTerragruntCommandWithOutput(t, "terragrunt run apply --backend-bootstrap --all --non-interactive --log-level debug --working-dir "+rootPath)
 	require.NoError(t, err)
